ARCH-1: extract mv dispatch arm into commands::mv::run (iter-225)
